Define `bats` upon first usage
I'm a software engineer that has used GitHub and Bash for years, but I had never heard of `bats` (seems excellent!). So, while reading this page in its current form, I had to chose between going to look up `bats`, which broke my flow of reading, or continuing in ignorance. I'm proposing this short definition of `bats`, with a link to the main GitHub repo for `bats`, to make reading this page flow more smoothly.